+++ generators/app/templates/demo.includes/widgets/example/my-editor-widget/spec/spec_runner.js
@@ -1,10 +1,12 @@
-<%- banner%>
+<%- banner %>
 ( function( global ) {
    'use strict';
    global.laxarSpec = {
       title: 'MyEditorWidget Specification',
       tests: [
          'my-editor-widget.spec'
-      ]
+      ],
+      testRunner: 'laxar-mocks',
+      jasmineMajorVersion: 2
    };
 } )( this );