+++ generators/widget/templates/spec/spec_runner.js
@@ -5,6 +5,8 @@
       title: '<%= name %> Specification',
       tests: [
          '<%= name %>.spec'
-      ]
-   };
+      ],
+      testRunner: 'laxar-mocks',
+      jasmineMajorVersion: 2
+  };
 } )( this );