+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Last Changes
 
+- [#1](https://github.com/LaxarJS/generator-laxarjs/issues/1): changed spec-tests to use laxar-mocks rather than laxar/laxar-testing
 - [#2](https://github.com/LaxarJS/generator-laxarjs/issues/2): created one configuration file for bower.json defaults for the scaffolds